Ensure that multiplication of clock offset is done on 64-bit
This is not a bug in UST per se, because we happen to force
multiplication by a ULL number, but force cast to uint64_t anyway to
ensure we don't trigger this overflow if the constant ever happen to
change.

+++ liblttng-ust/ltt-events.c
@@ -1088,7 +1088,7 @@ uint64_t measure_clock_offset(void)
 		return 0;
 	monotonic[1] = trace_clock_read64();
 	offset = (monotonic[0] + monotonic[1]) >> 1;
-	realtime = rts.tv_sec * 1000000000ULL;
+	realtime = (uint64_t) rts.tv_sec * 1000000000ULL;
 	realtime += rts.tv_nsec;
 	offset = realtime - offset;
 	return offset;
